pyproject.toml
[tool.poetry]
name = "toxicity-moderation"
version = "0.1.0"
description = ""
authors = ["Eren-Ajani Tshimanga <eren.tish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.43.1"
python-dotenv = "^1.0.1"
ollama = "^0.3.3"
fastapi = "^0.114.2"
requests = "^2.32.3"
pylint = "^3.3.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
